% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quartiles.R
\name{get_quartiles}
\alias{get_quartiles}
\title{Get Quartiles}
\usage{
get_quartiles(data, summarized_variable, grouping_variable = NULL)
}
\arguments{
\item{data}{the input data table to be analyzed. Named because data is the
typical thing to analyze in this fashion (and for consistency with dplyr)}

\item{summarized_variable}{the name of a numerical column in \code{data} for which
to calculate the quartiles. Named because this variable will be summarized.}

\item{grouping_variable}{(optional) the name of a column in \code{data} with a
categorical variable. If specified, the results will be summarized within
each unique entry of \code{grouping_variable}. Named since this is the variable
used for grouping the output.}
}
\value{
a tibble of showing the lower quartile, mean, and upper
quartile of \code{summarized_variable}, ignoring NA values. If \code{grouping_variable}
is not specified, the table will have one row representing the entire data. If
\code{grouping_variable} is specified, each row shows the quartile within one
unique category, which will be represented with an additional column.
}
\description{
This function returns a table of the quartiles for a given
numerical variable in a tibble, optionally divided into categories based on
an additional column.
}
\details{
this function utilizes \code{group_by()} and \code{summarize()} from the dplyr
package.
}
\examples{
get_quartiles(mtcars,wt) # Get the quartiles of weight over the whole dataset
get_quartiles(mtcars,wt,cyl) # Get weight quartiles divided by number of cylinders

}
